LayoutTests/http/tests/security/mixedContent/insecure-iframe-in-sandboxed-iframe.html

<!DOCTYPE html>
<body>
<script>
if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
    testRunner.dumpFrameLoadCallbacks();
}
</script>

<p>This test embeds a secure iframe which tries to open mixed content.
We should block mixed content even though the parent frame is insecure
because the middle frame is HTTPS.</p>

<script>
onload = function() {
    let ifr = document.createElement("iframe");
    ifr.sandbox = "allow-scripts";

    ifr.onload = function() {
        if (window.testRunner)
            testRunner.notifyDone();
    };
    ifr.src = "https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-iframe.html";

    document.body.appendChild(ifr);
};
</script>
</body>
